move sort functions to their own package
