+++ comp/metadata/host/utils/tags.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/config"
 	configUtils "github.com/DataDog/datadog-agent/pkg/config/utils"
-	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/util/cache"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/gce"
 	"github.com/DataDog/datadog-agent/pkg/util/docker"
@@ -22,6 +21,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/clustername"
 	k8s "github.com/DataDog/datadog-agent/pkg/util/kubernetes/hostinfo"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"github.com/DataDog/datadog-agent/pkg/util/sort"
 )
 
 var (
@@ -177,7 +177,7 @@ func GetHostTags(ctx context.Context, cached bool, conf config.Reader) *Tags {
 	}
 
 	t := &Tags{
-		System:              util.SortUniqInPlace(hostTags),
+		System:              sort.UniqInPlace(hostTags),
 		GoogleCloudPlatform: gceTags,
 	}
