+++ pkg/tagset/hashless_tags_accumulator.go
@@ -5,7 +5,7 @@
 
 package tagset
 
-import "github.com/DataDog/datadog-agent/pkg/util"
+import "github.com/DataDog/datadog-agent/pkg/util/sort"
 
 // HashlessTagsAccumulator allows to build a slice of tags, in a context where the hashes for
 // those tags are not useful.
@@ -59,7 +59,7 @@ func (h *HashlessTagsAccumulator) Copy() []string {
 
 // SortUniq sorts and remove duplicate in place
 func (h *HashlessTagsAccumulator) SortUniq() {
-	h.data = util.SortUniqInPlace(h.data)
+	h.data = sort.UniqInPlace(h.data)
 }
 
 // Reset resets the size of the builder to 0 without discarding the internal
