+++ pkg/util/cloudproviders/cloudproviders.go
@@ -12,9 +12,9 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/metadata/inventories"
-	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/util/kubelet"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"github.com/DataDog/datadog-agent/pkg/util/sort"
 
 	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 
@@ -141,7 +141,7 @@ func GetHostAliases(ctx context.Context) []string {
 	}
 	wg.Wait()
 
-	return util.SortUniqInPlace(aliases)
+	return sort.UniqInPlace(aliases)
 }
 
 // GetPublicIPv4 returns the public IPv4 from different providers
