+++ pkg/util/sort/sort_uniq.go
@@ -1,15 +1,15 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
+// Copyright 2023-present Datadog, Inc.
 
-package util
+package sort
 
 import "sort"
 
-// SortUniqInPlace sorts and remove duplicates from elements in place
+// UniqInPlace sorts and remove duplicates from elements in place
 // The returned slice is a subslice of elements
-func SortUniqInPlace(elements []string) []string {
+func UniqInPlace(elements []string) []string {
 	if len(elements) < 2 {
 		return elements
 	}
