+++ pkg/util/sort/sort_uniq_test.go
@@ -1,9 +1,11 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
+// Copyright 2023-present Datadog, Inc.
 
-package util
+//go:build test
+
+package sort
 
 import (
 	"fmt"
@@ -13,9 +15,9 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestSortUniqInPlace(t *testing.T) {
+func TestUniqInPlace(t *testing.T) {
 	elements := []string{"tag3:tagggg", "tag2:tagval", "tag1:tagval", "tag2:tagval"}
-	elements = SortUniqInPlace(elements)
+	elements = UniqInPlace(elements)
 
 	assert.ElementsMatch(t, elements, []string{"tag1:tagval", "tag2:tagval", "tag3:tagggg"})
 }
@@ -35,7 +37,7 @@ func benchmarkDeduplicateTags(b *testing.B, numberOfTags int) {
 
 	for n := 0; n < b.N; n++ {
 		copy(tempTags, tags)
-		SortUniqInPlace(tempTags)
+		UniqInPlace(tempTags)
 	}
 }
 func BenchmarkDeduplicateTags(b *testing.B) {
